requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
crypto==1.4.1
cryptography==38.0.1
dj-database-url==1.0.0
Django==4.1.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jwt==1.3.1
Naked==0.1.31
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.15.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
django-tinymce==3.5.0
tinycss2==1.1.1
